+++ AmpiriSDK.framework/Versions/A/Headers/AMPMediaContainerView.h
@@ -1,0 +1,14 @@
+//
+//  AMPMediaContainerView.h
+//  AmpiriSDK
+//
+//  Created by Glispa GmbH on 15/06/16.
+//  Copyright © 2016 Glispa GmbH. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface AMPMediaContainerView : UIView
+
+@property (assign, nonatomic) IBInspectable BOOL autoPlay;
+@end