+++ DemoApp/AmpiriSDKDemoApp/AMPDataUnitManager.h
@@ -1,0 +1,16 @@
+//
+//  AMPDataUnitManager.h
+//  AmpiriSDKTestApp
+//
+//  Created by Glispa GmbH on 20/06/16.
+//  Copyright © 2016 Glispa GmbH. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+#import <UIKit/UIKit.h>
+
+@interface AMPDataUnitManager : NSObject
+
++ (NSArray*)createDataUnitList;
+
+@end