+++ DemoApp/AmpiriSDKDemoApp/AppDelegate.m
@@ -16,6 +16,7 @@
 
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
+    [AmpiriSDK setTestMode:YES];
     return YES;
 }